requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.44.0
peft>=0.11.0
albumentations>=1.4.0
opencv-python>=4.9.0
pillow>=10.3.0
numpy>=1.26.0
pandas>=2.1.0
tqdm>=4.66.0
rich>=13.7.0
imagehash>=4.3.1
faiss-cpu>=1.7.4
scikit-learn>=1.4.0
seaborn>=0.13.0
matplotlib>=3.8.0
flake8>=7.0.0
black>=24.4.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pytest>=8.3.0
pytest-sugar>=0.9.7
httpx>=0.28.0
datasets>=2.20.0
